Switch Concurrency test to use catch2
- added verbose output to try to help catch an infrequent failure.
