+++ FWCore/Concurrency/test/BuildFile.xml
@@ -1,6 +1,9 @@
-<bin name="testFWCoreConcurrency" file="*.cppunit.cpp">
-  <use name="cppunit"/>
+<bin name="testFWCoreConcurrency" file="test2_catch2_*.cc">
+  <use name="catch2"/>
   <use name="FWCore/Concurrency"/>
+  <use name="tbb"/>
+  <!-- be verbose to help find failure-->
+  <flags TEST_RUNNER_CMD="testFWCoreConcurrency -s"/>
 </bin>
 
 <bin file="test_threadSafeOutputFileStream.cpp">
